Integrated a simple delay hook.
As discussed in the meeting on Friday, I implemented the delay hook for my simple model based on the Understanding Host Network Paper. It utilizes measurements of both C2M and P2M measured in the paper while varing the number of C2M cores. The default for the experiment has been set to 6 since that will induce the largest delay.